+++ packages/vike/src/plugins/index.ts
@@ -1,5 +1,6 @@
 import type { Plugin } from "vite";
 import { fixEnvsPlugins } from "./fix-envs";
 import { routesPlugin } from "./routes";
+import { buildAppPlugins } from "./build-app";
 
-export const plugins: Plugin[] = [fixEnvsPlugins(), routesPlugin()];
+export const plugins: Plugin[] = [fixEnvsPlugins(), routesPlugin(), buildAppPlugins()];
